refactor: fix compilation issues
