+++ sgxvm/src/querier.rs
@@ -1,4 +1,8 @@
+use std::vec::Vec;
 use crate::memory::{UnmanagedVector, U8SliceView};
+use crate::types::AllocationWithResult;
+use crate::ocall;
+use sgx_types::sgx_status_t;
 
 #[repr(C)]
 #[derive(Clone)]
@@ -23,4 +27,36 @@ pub struct Querier_vtable {
         *mut UnmanagedVector, // result output
         *mut UnmanagedVector, // error message output
     ) -> i32,
+}
+
+pub fn make_request(querier: *mut GoQuerier, request: Vec<u8>) -> Option<Vec<u8>> {
+    let mut allocation = std::mem::MaybeUninit::<AllocationWithResult>::uninit();
+
+    let result = unsafe {
+        ocall::ocall_query_raw(
+            allocation.as_mut_ptr(),
+            querier,
+            request.as_ptr(),
+            request.len(),
+        )
+    };
+
+    match result {
+        sgx_status_t::SGX_SUCCESS => {
+            let allocation = unsafe { allocation.assume_init() };
+            let result_vec = unsafe {
+                Vec::from_raw_parts(
+                    allocation.result_ptr,
+                    allocation.result_len,
+                    allocation.result_len,
+                )
+            };
+
+            return Some(result_vec);
+        }
+        _ => {
+            println!("make_request failed: Reason: {:?}", result.as_str());
+            return None;
+        }
+    };
 }
